模型整改

+++ micro/client/controller/article.go
@@ -16,7 +16,7 @@ import (
 )
 
 type Article struct {
-	ID            uint             `gorm:"primary_key" json:"id"`
+	ID            uint64           `gorm:"primary_key" json:"id"`
 	CreatedAt     time.Time        `json:"created_at"`
 	Title         string           `gorm:"type:varchar(100)" json:"title"`
 	Intro         string           `gorm:"type:varchar(100)" json:"intro"`
@@ -28,22 +28,22 @@ type Article struct {
 	Categories    []Category       `gorm:"many2many:article_category" json:"categories"` //分类
 	Tags          []Tag            `gorm:"many2many:article_tag;foreignkey:ID;association_foreignkey:Name" json:"tags"`
 	User          User             `json:"user"`
-	UserID        uint             `json:"user_id"`
+	UserID        uint64           `json:"user_id"`
 	Comments      []ArticleComment `json:"comments"`                                  //评论
-	BrowseCount   uint             `json:"browse_count"`                              //浏览
-	CommentCount  uint             `gorm:"default:0" json:"comment_count"`            //评论
-	CollectCount  uint             `gorm:"default:0" json:"collect_count"`            //收藏
-	ApproveCount  uint             `gorm:"default:0" json:"approve_count"`            //点赞
-	LikeCount     uint             `gorm:"default:0" json:"like_count"`               //喜欢
+	BrowseCount   uint64           `json:"browse_count"`                              //浏览
+	CommentCount  uint64           `gorm:"default:0" json:"comment_count"`            //评论
+	CollectCount  uint64           `gorm:"default:0" json:"collect_count"`            //收藏
+	ApproveCount  uint64           `gorm:"default:0" json:"approve_count"`            //点赞
+	LikeCount     uint64           `gorm:"default:0" json:"like_count"`               //喜欢
 	Permission    int8             `gorm:"type:smallint;default:0" json:"permission"` //查看权限
 	Sequence      int8             `gorm:"type:smallint;default:0" json:"sequence"`   //排序，置顶
 	UpdatedAt     *time.Time       `json:"updated_at"`
 	DeletedAt     *time.Time       `sql:"index" json:"deleted_at"`
-	Status        uint             `json:"status"`                        //状态
-	ModifyTimes   uint             `gorm:"default:0" json:"modify_times"` //修改次数
-	ParentID      uint             `json:"parent_id"`                     //父ID
+	Status        uint64           `json:"status"`                        //状态
+	ModifyTimes   uint64           `gorm:"default:0" json:"modify_times"` //修改次数
+	ParentID      uint64           `json:"parent_id"`                     //父ID
 	LastUser      User             `json:"last_user"`
-	LastUserID    uint             `json:"last_user_id"` //最后一个回复话题的人
+	LastUserID    uint64           `json:"last_user_id"` //最后一个回复话题的人
 	LastCommentAt *time.Time       `json:"last_comment_at"`
 }
 
@@ -309,7 +309,7 @@ func AddArticle(c iris.Context) {
 	common.Response(c, "保存成功", e.SUCCESS)
 }
 
-func historyArticle(c iris.Context, isDel uint) (model.ArticleHistory, model.Article, error) {
+func historyArticle(c iris.Context, isDel uint8) (model.ArticleHistory, model.Article, error) {
 
 	var article model.Article
 	//获取文章ID
