+++ micro/client/controller/upload/upload.go
@@ -135,7 +135,7 @@ func GetDirAndUrl(classify string, info *multipart.FileHeader) (string, string, 
 
 // Upload 文件上传
 func Upload(ctx iris.Context) *model.FileUploadInfo {
-	userId := ctx.Values().Get("userId").(uint)
+	userId := ctx.Values().Get("userId").(uint64)
 	classify := ctx.Params().GetString("classify")
 	file, info, err := ctx.FormFile("file")
 	md5 := ctx.FormValue("md5")
@@ -178,7 +178,7 @@ func Upload(ctx iris.Context) *model.FileUploadInfo {
 		return nil
 	}
 
-	upInfo.File.Size = uint(info.Size)
+	upInfo.File.Size = uint64(info.Size)
 	upInfo.UploadUserID = userId
 	upInfo.Status = 1
 	upInfo.MD5 = md5
@@ -215,8 +215,8 @@ func UploadMultiple(ctx iris.Context) {
 			failures++
 			common.Response(ctx, file[0].Filename+"上传失败")
 		} else {
-			upInfo.File.Size = uint(file[0].Size)
-			userId := ctx.Values().Get("userId").(uint)
+			upInfo.File.Size = uint64(file[0].Size)
+			userId := ctx.Values().Get("userId").(uint64)
 			upInfo.UploadUserID = userId
 			if err := initialize.DB.Create(&upInfo).Error; err != nil {
 				common.Response(ctx, err.Error())
@@ -263,7 +263,7 @@ func SaveUploadedFile(file *multipart.FileHeader, dir string, url string) (*mode
 }
 
 func MD5(ctx iris.Context) {
-	userId := ctx.Values().Get("userId").(uint)
+	userId := ctx.Values().Get("userId").(uint64)
 	md5 := ctx.Params().Get("md5")
 	var upI model.FileUploadInfo
 	var count int
