+++ micro/model/collection.go
@@ -5,13 +5,13 @@ import (
 )
 
 type Favorites struct {
-	ID          uint         `gorm:"primary_key" json:"id"`
+	ID          uint64       `gorm:"primary_key" json:"id"`
 	CreatedAt   time.Time    `json:"created_at"`
 	Name        string       `gorm:"type:varchar(20)" json:"name"`
 	User        User         `json:"user"`
-	UserID      uint         `json:"user_id"`
+	UserID      uint64       `json:"user_id"`
 	FollowUsers []User       `json:"follow_users"`
-	Count       uint         `json:"count"`
+	Count       uint64       `json:"count"`
 	Collections []Collection `gorm:"many2many:collection_favorites" json:"collections"`
 	UpdatedAt   *time.Time   `json:"updated_at"`
 	DeletedAt   *time.Time   `sql:"index" json:"deleted_at"`
@@ -20,32 +20,32 @@ type Favorites struct {
 
 //收藏夹？像网易云一样可以收藏别人的歌单
 type Collection struct {
-	ID          uint        `gorm:"primary_key" json:"id"`
+	ID          uint64      `gorm:"primary_key" json:"id"`
 	CreatedAt   time.Time   `json:"created_at"`
-	RefID       uint        `json:"ref_id"`
+	RefID       uint64      `json:"ref_id"`
 	Kind        string      `gorm:"type:varchar(10)" json:"kind"`
 	Favorites   []Favorites `json:"favorites"`
-	FavoritesID uint        `json:"favorites_id"`
-	UserID      uint        `json:"user_id"`
+	FavoritesID uint64      `json:"favorites_id"`
+	UserID      uint64      `json:"user_id"`
 	UpdatedAt   *time.Time  `json:"updated_at"`
 	DeletedAt   *time.Time  `sql:"index" json:"deleted_at"`
 	Status      uint8       `gorm:"type:smallint;default:0" json:"status"`
 }
 
 type Like struct {
-	ID        uint       `gorm:"primary_key" json:"id"`
+	ID        uint64     `gorm:"primary_key" json:"id"`
 	CreatedAt time.Time  `json:"created_at"`
-	RefID     uint       `json:"ref_id"`
+	RefID     uint64     `json:"ref_id"`
 	Kind      string     `gorm:"type:varchar(10)" json:"kind"`
 	User      User       `json:"user"`
-	UserID    uint       `json:"user_id"`
-	Count     uint       `json:"count"`
+	UserID    uint64     `json:"user_id"`
+	Count     uint64     `json:"count"`
 	UpdatedAt *time.Time `json:"updated_at"`
 	DeletedAt *time.Time `sql:"index" json:"deleted_at"`
 	Status    uint8      `gorm:"type:smallint;default:0" json:"status"`
 }
 
 type ArticleCollection struct {
-	ArticleID   uint         `json:"article_id"`
+	ArticleID   uint64       `json:"article_id"`
 	Collections []Collection `json:"collections"`
 }
